% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runappserver.R
\name{runQuestManualAppServer}
\alias{runQuestManualAppServer}
\title{Start the Survey Solutions Questionnaire Manual Application on Shiny Server}
\usage{
runQuestManualAppServer()
}
\description{
A wrappter function to start the application on the SERVER. Please make sure you have read the
documentation.
}
\details{
This function only works on shiny server to do so, create a script in the app directory and
run this function. An example myListingApp.R file could be:

\if{html}{\out{<div class="sourceCode">}}\preformatted{susoquestionnairemanual::runQuestManualApp()

}\if{html}{\out{</div>}}
}
