% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susoquestionnairemanual-package.R
\docType{package}
\name{susoquestionnairemanual-package}
\alias{susoquestionnairemanual}
\alias{susoquestionnairemanual-package}
\title{susoquestionnairemanual: Shiny Application to generate a questionnaire manual from Survey Solutions Questionnaire}
\description{
A simple shiny application to generate a questionnaire document in either HTML, MS Word, or in MS Powerpoint. Questionnaire can be uploaded as the exported json, or directly from the users Survey Solutions Server.
}
\author{
\strong{Maintainer}: Michael Wild \email{mwild@worldbank.org}

}
\keyword{internal}
